add base sql
