+++ src/test/mod.rs
@@ -1,2 +1,3 @@
+pub mod base_sql;
 pub mod sled;
 pub mod test_util;
