+++ src/test/test_util.rs
@@ -9,7 +9,10 @@ use uuid::Uuid;
 
 pub async fn create_execution() -> MysqlResult<Execution> {
     let mut my_config = MyConfig::default();
-    let test_id = Uuid::new_v4().to_simple().encode_lower(&mut Uuid::encode_buffer()).to_string();
+    let test_id = Uuid::new_v4()
+        .to_simple()
+        .encode_lower(&mut Uuid::encode_buffer())
+        .to_string();
     my_config.engine.sled.data_path = format!("./data/ebike/sled/{}", test_id);
 
     let global_context = Arc::new(Mutex::new(GlobalContext::new_with_config(
@@ -79,43 +82,4 @@ pub async fn create_execution() -> MysqlResult<Execution> {
     }
 
     Ok(core_execution)
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::core::output::CoreOutput;
-    use crate::core::output::CoreOutput::FinalCount;
-    use crate::mysql::error::MysqlResult;
-    use crate::mysql::{message, metadata};
-    use crate::test::test_util::create_execution;
-    use arrow::datatypes::{DataType, Field, Schema};
-    use arrow::record_batch::RecordBatch;
-    use datafusion::{assert_batches_eq, assert_batches_sorted_eq};
-    use std::sync::Arc;
-
-    #[tokio::test]
-    async fn show_databases() -> MysqlResult<()> {
-        let mut core_execution = create_execution().await?;
-
-        let result = core_execution.execute_query("show databases").await?;
-
-        let mut results: Vec<RecordBatch> = vec![];
-        match result {
-            CoreOutput::ResultSet(_, r) => results = r,
-            _ => {}
-        }
-
-        let expected = vec![
-            "+--------------------+",
-            "| Database           |",
-            "+--------------------+",
-            "| mysql              |",
-            "| performance_schema |",
-            "+--------------------+",
-        ];
-
-        assert_batches_eq!(expected, &results);
-
-        Ok(())
-    }
 }
